requirements.txt
gradio==4.28.3
langchain_community==0.0.36
langchain_core==0.1.50
numpy==1.25.0
opencv_contrib_python==4.7.0.72
opencv_python==4.8.1.78
opencv_python_headless==4.8.0.76
Pillow==10.3.0
torch==2.0.0+cu118
